requirement.txt.txt
Flask==2.1.0
mistletoe==0.9.1
google-generativeai==0.0.5
